chore: rename context files to jsx
