+++ admin.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AdminProvider } from './components/AdminContext';
+import { AdminProvider } from './components/AdminContext.jsx';
 import AdminApp from './components/AdminApp';
 
 export default function Admin() {
