+++ home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AppProvider } from './components/context';
+import { AppProvider } from './components/context.jsx';
 import CharismaMoveApp from './components/CharismaMoveApp';
 
 export default function App() {
